added flag to hits that we expect SOCK_STREAM address
This is to fix issue when varnish (3.0.3) would be serving garbled
status line for the first served response when configured with backend
probes on Max OS. This was caused by getaddrinfo call itself and went
away whan this flag was added.
I cannot explain what is the root cause of this strange behaviour though.